Refactor loadParameterClasses to RedisUtils and update tests
Moved the loadParameterClasses utility from RedisCommandsUtils to RedisUtils for better reuse. Updated all references and related tests to use the new location, and removed the redundant implementation from RedisCommandsUtils.
