+++ microsphere-redis-core/src/test/java/io/microsphere/redis/util/RedisUtilsTest.java
@@ -25,8 +25,11 @@ import java.util.Set;
 import static io.microsphere.redis.util.RedisCommandUtils.LOAD_REDIS_COMMANDS_FUNCTION;
 import static io.microsphere.redis.util.RedisCommandUtils.REDIS_COMMANDS_RESOURCE;
 import static io.microsphere.redis.util.RedisCommandUtils.getRedisCommands;
+import static io.microsphere.redis.util.RedisUtils.loadParameterClasses;
 import static io.microsphere.redis.util.RedisUtils.loadResources;
 import static java.util.Collections.emptySet;
+import static java.util.stream.Stream.of;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
@@ -46,5 +49,18 @@ class RedisUtilsTest {
         assertEquals(redisCommands, allRedisCommands);
 
         assertEquals(emptySet(), loadResources("REDIS_COMMANDS_RESOURCE", inputStreams -> emptySet()));
+    }
+
+    @Test
+    void testLoadParameterClasses() {
+        assertLoadParameterClasses();
+        assertLoadParameterClasses(String.class);
+        assertLoadParameterClasses(String.class, Integer.class);
+        assertLoadParameterClasses(String.class, Integer.class, RawValue.class);
+        assertLoadParameterClasses(String.class, Integer.class, RawValue.class, this.getClass());
+    }
+
+    private void assertLoadParameterClasses(Class<?>... classes) {
+        assertArrayEquals(classes, loadParameterClasses(of(classes).map(Class::getName).toArray(String[]::new)));
     }
 }
