+++ microsphere-redis-spring/src/test/java/io/microsphere/redis/spring/util/RedisCommandsUtilsTest.java
@@ -18,15 +18,13 @@
 package io.microsphere.redis.spring.util;
 
 
+import io.microsphere.redis.metadata.ParameterMetadata;
 import io.microsphere.redis.spring.config.RedisConfig;
 import io.microsphere.redis.spring.config.RedisContextConfig;
 import io.microsphere.redis.spring.context.RedisContext;
 import io.microsphere.redis.spring.event.RedisCommandEvent;
 import io.microsphere.redis.spring.interceptor.RedisMethodContext;
-import io.microsphere.redis.metadata.ParameterMetadata;
-import io.microsphere.redis.util.RedisCommandUtils;
 import org.junit.jupiter.api.Test;
-import org.springframework.context.ApplicationContext;
 import org.springframework.data.redis.connection.RedisConnection;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.RedisKeyCommands;
@@ -56,14 +54,12 @@ import static io.microsphere.redis.spring.util.RedisCommandsUtils.REDIS_ZSET_COM
 import static io.microsphere.redis.spring.util.RedisCommandsUtils.buildCommandMethodId;
 import static io.microsphere.redis.spring.util.RedisCommandsUtils.getRedisCommands;
 import static io.microsphere.redis.spring.util.RedisCommandsUtils.initializeParameters;
-import static io.microsphere.redis.spring.util.RedisCommandsUtils.loadParameterClasses;
 import static io.microsphere.redis.spring.util.RedisCommandsUtils.resolveInterfaceName;
 import static io.microsphere.redis.spring.util.RedisCommandsUtils.resolveSimpleInterfaceName;
 import static io.microsphere.redis.util.RedisCommandUtils.buildMethodId;
 import static io.microsphere.reflect.MethodUtils.findMethod;
 import static io.microsphere.spring.test.util.SpringTestUtils.testInSpringContainer;
 import static io.microsphere.util.ArrayUtils.EMPTY_OBJECT_ARRAY;
-import static java.util.stream.Stream.of;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -195,19 +191,6 @@ class RedisCommandsUtilsTest {
 
         assertFalse(initializeParameters(randomKeyMethod, EMPTY_OBJECT_ARRAY, (parameter, integer) -> {
         }));
-    }
-
-    @Test
-    void testLoadParameterClasses() {
-        assertLoadParameterClasses();
-        assertLoadParameterClasses(String.class);
-        assertLoadParameterClasses(String.class, Integer.class);
-        assertLoadParameterClasses(String.class, Integer.class, ApplicationContext.class);
-        assertLoadParameterClasses(String.class, Integer.class, ApplicationContext.class, this.getClass());
-    }
-
-    private void assertLoadParameterClasses(Class<?>... classes) {
-        assertArrayEquals(classes, loadParameterClasses(of(classes).map(Class::getName).toArray(String[]::new)));
     }
 
     void assertInterfaceName(String interfaceName) {
